requirements.txt
Pillow~=7.1.2
pip~=20.1
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
MarkupSafe~=1.1.1
setuptools~=46.3.0
Flask~=1.1.2